Rename interfaces for consistent capitalization
