+++ src/_models/SystemState.ts
